requirements.txt
beautifulsoup4==4.8.2
aiofiles==0.4.0
aiohttp==3.6.2
asyncio==3.4.3
asyncpg==0.20.1
git+git://github.com/MujyKun/boats.py@master#boats.py
dblpy==0.4.0
discord.py==1.5.1
tweepy==3.8.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
PyNaCl==1.3.0
pypapago==0.1.1.1
youtube-dl==2020.9.20
python-dotenv==0.14.0
filetype==1.0.7
xmltodict==0.12.0
ksoftapi==0.2.4a0
Pillow==8.0.1
datadog==0.39.0
git+git://github.com/MujyKun/Weverse@main#Weverse
pytz==2020.4
parsedatetime==2.6
Flask==1.1.1
pytest==6.2.1
pytest-asyncio==0.14.0
mock==4.0.3